Adding three fields to clockwork.toml
These fields are going to be used for calling `sacct` through SSH. These fields are already in the `hosts.yml` but they don't make it to `/etc/clockwork/clockwork.toml` at the moment and that means that we cannot use them. This commit will fix that.

This related to cw204 branch of Clockwork, and in turn to `https://github.com/mila-iqia/clockwork/pull/88`.