lowercase for boolean values
There's a problem at the moment with the fact that the boolean value "cluster.sacct_enabled" gets written with an uppercase in the toml scripts, and that's not even valid toml. This can be fixed with the ajustment of adding a `|lower` in the jinja2 template.

The consequence of this problem is that the deployment on `clockwork2-dev` is broken and needs to be adjusted manually by editing `/etc/clockwork/clockwork.toml`, which is not an acceptable solution. For this reason, and because this Ansible recipe is not used anywhere else, it's relatively safe to merge this into master and we could always backpedal if things break in practice (they're broken right now anyways).